+++ cpp/src/arrow/python/config.cc
