+++ cpp/src/arrow/python/helpers.h
