+++ cpp/src/arrow/python/io.cc
