+++ cpp/src/arrow/python/io.h
