+++ cpp/src/arrow/python/type_traits.h
